feat(bond): Add bond manager
When a Sequencer wants to appendBatch, they need to have previously
called BondManager.deposit in order to have stake available to be locked.

Inside Transitioners, whenever somebody provides a witness, their contribution
for that dispute (keyed by pre-state root) is stored. Later on, once the dispute
is resolved, the fraud verifier makes a call to the BondManager and if there was
no fraud, then the sequencer's bond is returned, otherwise it is distributed
proportionally to all the witness providers that helped resolve that dispute