feat(fraud-verifier): reward callers of initialize/finalizeFraudVerification
NOTE: This requires that the address of the BondManager is already set in the AddressManager. However, since the BondManager also needs the address of the FraudVerifier, this creates a cyclic dependency. To solve it, you MUST pre-calculate the address of both contracts before deployment (by hashing the deployer address + nonce) and set that in the AddressManager